+++ quantify_core/data/handling.py
@@ -772,7 +772,10 @@ def snapshot(update: bool = False, clean: bool = True) -> OrderedDict:
         }
     )
     for ins_name, ins_ref in Instrument._all_instruments.items():
-        snap["instruments"][ins_name] = ins_ref().snapshot(update=update)
+        ref = ins_ref()
+        # Check for dead weakrefs
+        if ref is not None:
+            snap["instruments"][ins_name] = ref.snapshot(update=update)
 
     if clean:
         exclude_keys = {