+++ CHANGELOG.rst
@@ -14,8 +14,7 @@ Unreleased
 * Bugfix - Rabi analysis for negative signal amplitudes can now converge.
 * Docs - Added bibliography with sphinxcontrib-bibtex extension (!207).
 * Bugfix - Fixed divide by 0 warning in resonator spectroscopy analysis (!216).
-
-
+* Bugfix - Fixed snapshot failing for qcodes instruments with dead weakrefs (!221).
 
 0.5.0 (2021-08-06)
 ------------------